feat(web): extract TracesContent and MobileContent from page components — SP 2.3
TracesContent accepts projectId + optional traceId props, contains all trace
list rendering logic and tRPC query. MobileContent accepts projectId prop with
null defaults for deep-link params. traces/page.tsx refactored to delegate to
TracesContent. mobile/page.tsx intentionally not refactored (deep-link params
make delegation impractical without interface changes).
